implement signing up by sending us an email
this is preferred over putting an email address in the web form: by letting the
user send us the first message, our address will be a known correpondent in
their email account. that should help their junk filter realize our signup
message and later notification messages are legit. it also makes it impossible
for users to mistype their own email. we make it easy for users: they just
click our link to compose a message and hit send in their email application.

we'll now monitor the inbox for messages with "signup" as subject.
some metrics related to incoming message handling were changed.
messages are only processed if they don't have headers that indicate they are
coming from a mailing list or are automated in any way. we also check are
either have a dmarc pass, or aligned relaxed spf or aligned relaxed dkim pass,
based on the authentication-results header the mail server must add to the
message.

this adds config options to disable signup through email or website (or both to
disable entirely).

this changes config option SubjectPrefix. it must now be the full prefix,
no ": " is added anymore.

this also increases likelyhood of being able to deliver messages to address
with non-ascii localparts by setting the smtputf8 & 8bitmime extension when
needed.

for issue #25