add support for the new webapi/webhooks in mox
as alternative to composing messages ourselves, submitting them over smtp with
the smtpclient package, and monitoring an imap mailbox for dsns with the
imapclient package. now it's just http/json calls.

the previous compose/imap/submission is kept, to run this without mox, and to
show the difference between the two.