fix certain dnssec responses, with the right nsec type bitmap and not sending a nxname response for the bare v0 name
for v0.l.gopherwatch.org we were giving back "nxname", which would mean there's nothing below.

but this wasn't what was causing the problem with quad9 resolvers.

in the nsec type bitmap we were claiming records that didn't exist, i suspect
it's mostly the ns & dnskey that was giving trouble, but didn't test it.

with this change, i no longer see servfail from 9.9.9.9 resolvers (before it
was working randomly for their resolvers, possibly due to parallel requests,
and the order the responses would come back in).

for issue #29